Adding shortcodes for figures and images
I've added two shortcodes, taking inspiration from [Laura Kalbag's
blog](https://laurakalbag.com/processing-responsive-images-with-hugo/).
One is a shortcode for images which deals with automatic resizing and
responsive images. The other is a modification of the default figure
shortcode to be a bit more flexible. Check out
[/posts/arithmetic-algebra-games](/content/en/posts/arithmetic-algebra-games/index.md)
to see them in use.

Signed-off-by: Ian Allison <iana@pims.math.ca>